Chore: Update review content - API 이름 수정 - Getter,Setter -> Data - 예외 처리 try/catch -> MemberException - 폴더 구조 변경
