+++ src/main/java/hufs/likelion/gov/domain/authentication/entity/RefreshToken.java
@@ -1,4 +1,4 @@
-package hufs.likelion.gov.entity;
+package hufs.likelion.gov.domain.authentication.entity;
 
 import java.util.Date;
 
@@ -10,12 +10,10 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 
 @Entity
-@Getter
-@Setter
+@Data
 public class RefreshToken {
 
 	@Id
