+++ src/main/java/hufs/likelion/gov/domain/authentication/exception/MemberException.java
@@ -1,0 +1,12 @@
+package hufs.likelion.gov.domain.authentication.exception;
+
+public class MemberException extends RuntimeException {
+
+	public MemberException(String message) {
+		super(message);
+	}
+
+	public MemberException(String message, Throwable cause) {
+		super(message, cause);
+	}
+}
