+++ src/main/java/hufs/likelion/gov/domain/authentication/repository/MemberRepository.java
@@ -1,10 +1,10 @@
-package hufs.likelion.gov.repository;
+package hufs.likelion.gov.domain.authentication.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import hufs.likelion.gov.entity.Member;
+import hufs.likelion.gov.domain.authentication.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 	Optional<Member> findByMemberId(String memberId);
