+++ src/main/java/hufs/likelion/gov/domain/authentication/repository/RefreshTokenRepository.java
@@ -1,11 +1,11 @@
-package hufs.likelion.gov.repository;
+package hufs.likelion.gov.domain.authentication.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import hufs.likelion.gov.entity.Member;
-import hufs.likelion.gov.entity.RefreshToken;
+import hufs.likelion.gov.domain.authentication.entity.RefreshToken;
+import hufs.likelion.gov.domain.authentication.entity.Member;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
 	Optional<RefreshToken> findByToken(String token);
